server/requirements.txt
wheel
Flask
Flask-HTTPAuth
Flask-SQLAlchemy
Flask-JWT
Flask-Testing
Jinja2
MarkupSafe
SQLAlchemy
sqlalchemy-redshift
SQLAlchemy-migrate
itsdangerous
passlib
flask_script
pymysql

flask_socketio
gevent
scrypt

mysqlclient
requests
socketIO_client

python-slugify

#psycopg2
